requirements.txt
PyYAML==3.09
pymongo>=2
gunicorn==0.12.0
Flask>=0.7
Jinja2==2.4.1
BeautifulSoup
requests
python-googlegeocoder
nose
#python-nose
Django==1.4.0

# Also need to install PyLevelDB
